Cross-reference post GUID fetch by blogID
This allows us to add two feeds, say https://www.joanwestenberg.com/feed/ and https://www.joanwestenberg.com/rss/, which have different final URLs but the same sets of post GUIDs. Two lists, one with /feed/ and one with /rss/, will both be able to pull the blog. It does mean we end up with post duplicates, but that's an acceptable price to pay.